+++ CHANGELOG.md
@@ -13,6 +13,10 @@ use patch releases for compatibility fixes instead.
 
 - Added missing `url` field to `DatasetStorage`.
 
+### Removed
+
+- Removed old field `fileheap` from `DatasetSpec`.
+
 ## [v1.11.5](https://github.com/allenai/beaker-py/releases/tag/v1.11.5) - 2022-11-16
 
 ### Added